First iteration of Bogie Controller test script
Script takes 3 command line bytes (in decimal) and sends a packet over
serial with address, speed, and turning position set from the arguments.
I haven't had a chance to test it on the bb yet.  Also, it's set to use
tty01 but I'm not sure that's the correct port for the bogies.